refactoring AddUsers to InviteLearners, switch to revoked tab after success
