+++ src/containers/InviteLearnersModal/InviteLearnersModal.test.jsx
@@ -6,8 +6,8 @@ import configureMockStore from 'redux-mock-store';
 import thunk from 'redux-thunk';
 import { mount } from 'enzyme';
 
-import AddUsersModal from './index';
-import subscribeEmailTemplate from '../../components/AddUsersModal/emailTemplate';
+import InviteLearnersModal from './index';
+import subscribeEmailTemplate from '../../components/InviteLearnersModal/emailTemplate';
 
 const mockStore = configureMockStore([thunk]);
 
@@ -33,10 +33,10 @@ const initialState = {
   },
 };
 
-const UserSubscriptionModalWrapper = props => (
+const InviteLearnersModalWrapper = props => (
   <MemoryRouter>
     <Provider store={props.store}>
-      <AddUsersModal
+      <InviteLearnersModal
         title={modalTitle}
         availableSubscriptionCount={10}
         onClose={() => {}}
@@ -48,11 +48,11 @@ const UserSubscriptionModalWrapper = props => (
   </MemoryRouter>
 );
 
-UserSubscriptionModalWrapper.defaultProps = {
+InviteLearnersModalWrapper.defaultProps = {
   store: mockStore(initialState),
 };
 
-UserSubscriptionModalWrapper.propTypes = {
+InviteLearnersModalWrapper.propTypes = {
   store: PropTypes.shape({}),
 };
 
@@ -66,7 +66,7 @@ describe('UserSubscriptionModalWrapper', () => {
   });
 
   it('renders user licenses modal', () => {
-    const wrapper = mount(<UserSubscriptionModalWrapper data={data} />);
+    const wrapper = mount(<InviteLearnersModalWrapper data={data} />);
     expect(wrapper.find('.modal-title').text()).toEqual(modalTitle);
     expect(wrapper.find('.modal-body form h3').first().text()).toEqual('Add Users');
   });
